Rearrange Changelog module
